requirements.txt
loguru == 0.7.2
requests == 2.31.0
beautifulsoup4 == 4.12.2
tqdm == 4.66.2
python-dotenv == 1.0.1
psycopg2 == 2.9.9
